requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-cloud-bigquery
google-cloud-storage
firebase_admin
pandas
lfapi